+++ src/r3f/index.jsx
@@ -1,4 +1,4 @@
-export * from './TilesAttributionComponent.jsx';
+export * from './TilesRenderer.jsx';
+export * from './TilesAttributionOverlay.jsx';
 export * from './CanvasDOMOverlay.jsx';
 export * from './CameraControls.jsx';
-export * from './TilesAttributionComponent.jsx';
